indian-paper-trading-system/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
pytz>=2021.1
python-dotenv>=0.19.0
requests>=2.25.1
schedule>=1.1.0
finta>=1.3.0
holidays>=0.13
gymnasium>=0.28.1
stable-baselines3>=2.0.0
torch>=1.9.0
dash>=2.8.0
plotly>=5.10.0
tabulate>=0.9.0
rich>=12.0.0
kiteconnect>=4.0.0
dash-bootstrap-components>=1.2.0
tqdm>=4.62.0